+++ vhost/src/main.cpp
@@ -2,6 +2,8 @@
 
 #include "config.h"
 
+#include <rawstd/exitcode.h>
+
 #include <getopt.h>
 #include <signal.h>
 
@@ -9,9 +11,11 @@
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
+#include <sysexits.h>
 
 #include <iostream>
 #include <sstream>
+#include <system_error>
 
 #define DEFAULT_QUEUE_SIZE 256
 
@@ -105,7 +109,7 @@ int main(int argc, char** argv) {
             return EXIT_SUCCESS;
 
         default:
-            return EXIT_FAILURE;
+            return EX_USAGE;
         }
     }
 
@@ -116,7 +120,7 @@ int main(int argc, char** argv) {
 
     if (optind < argc) {
         std::cerr << "Unexpected argument: " << argv[optind] << std::endl;
-        return EXIT_FAILURE;
+        return EX_USAGE;
     }
 
     unsigned int queue_size = DEFAULT_QUEUE_SIZE;
@@ -125,18 +129,18 @@ int main(int argc, char** argv) {
         if (iss.peek() < '0' || iss.peek() > '9' || !(iss >> queue_size) ||
             !iss.eof()) {
             std::cerr << "queue-size must be unsigned integer" << std::endl;
-            return EXIT_FAILURE;
+            return EX_USAGE;
         }
     }
 
     if (socket_path_arg == nullptr) {
         std::cerr << "socket-path argument required" << std::endl;
-        return EXIT_FAILURE;
+        return EX_USAGE;
     }
 
     if (target_arg == nullptr) {
         std::cerr << "target argument required" << std::endl;
-        return EXIT_FAILURE;
+        return EX_USAGE;
     }
 
     sact.sa_handler = sact_handler;
@@ -146,21 +150,24 @@ int main(int argc, char** argv) {
         errno = 0;
         std::cerr << "Failed to register SIGINT handler: " << strerror(errsv)
                   << std::endl;
-        return EXIT_FAILURE;
+        return rawstd_exitcode_for_errno(errsv);
     }
     if (sigaction(SIGTERM, &sact, nullptr) == -1) {
         int errsv = errno;
         errno = 0;
         std::cerr << "Failed to register SIGTERM handler: " << strerror(errsv)
                   << std::endl;
-        return EXIT_FAILURE;
+        return rawstd_exitcode_for_errno(errsv);
     }
 
     try {
         server(queue_size, target_arg, socket_path_arg);
+    } catch (const std::system_error& e) {
+        std::cerr << e.what() << std::endl;
+        return rawstd_exitcode_for_errno(e.code().value());
     } catch (const std::exception& e) {
         std::cerr << e.what() << std::endl;
-        return EXIT_FAILURE;
+        return EX_SOFTWARE;
     }
 
     return EXIT_SUCCESS;